Move help text logic to a hook.
